fix: Fix Option nullable annotations
